target/riscv: Fix incorrect MIP update logic
At present the riscv_cpu_update_mip() logic has:

    if (new && !old) {
        cpu_interrupt(CPU(cpu), CPU_INTERRUPT_HARD);
    } else if (!new && old) {
        cpu_reset_interrupt(CPU(cpu), CPU_INTERRUPT_HARD);
    }

Considering the following situation:

    new = MTIP | MSIP
    old = MTIP

No IPI interrupt will be fired as it does not fall into the logic
of if (new && !old).

With this fix, the IPI is fully working for PLICSW on Andes AE350
platform with a 4-core configuration, tested by booting VxWorks 7.

Signed-off-by: Bin Meng <bin.meng@windriver.com>